+++ WANetworkRouting/AppDelegate.m
@@ -95,7 +95,8 @@
     WANetworkRoutingManager *routingManager = [WANetworkRoutingManager managerWithBaseURL:[NSURL URLWithString:@"http://jsonplaceholder.typicode.com"]
                                                                            requestManager:requestManager
                                                                            mappingManager:mappingManager
-                                                                    authenticationManager:nil];
+                                                                    authenticationManager:nil
+                                                                             batchManager:nil];
     
     // ———————————————————————————————————————————
     // Configure the router