+++ WANetworkRoutingTests/WARouteTests.m
@@ -16,7 +16,7 @@
 SPEC_BEGIN(WARouteTest)
 
 describe(@"Register a simple route", ^{
-    WANetworkRoutingManager *objectManager = [WANetworkRoutingManager managerWithBaseURL:[NSURL URLWithString:@"http://www.someURL.com"] requestManager:nil mappingManager:nil authenticationManager:nil];
+    WANetworkRoutingManager *objectManager = [WANetworkRoutingManager managerWithBaseURL:[NSURL URLWithString:@"http://www.someURL.com"] requestManager:nil mappingManager:nil authenticationManager:nil batchManager:nil];
     WANetworkRoute *routeGetAllEnterprises = [WANetworkRoute routeWithObjectClass:[Enterprise class]
                                                                       pathPattern:@"enterprises"
                                                                            method:WAObjectRequestMethodGET];